Move networkSet functionality to middleware
